% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{append_fasta_to_gff}
\alias{append_fasta_to_gff}
\title{Append a fasta entry to a GFF file}
\usage{
append_fasta_to_gff(gff_path)
}
\arguments{
\item{gff_path}{Character. Path to GFF file.}
}
\description{
Some incorrectly formatted GFF files have a fasta entry appended to the end.
This file generates and saves such an improper GFF file.
}
