% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_tree_and_fasta.R
\name{build_tree}
\alias{build_tree}
\title{Build tree}
\usage{
build_tree(mat)
}
\arguments{
\item{mat}{allele matrix where rows are samples and columns are variants
(generated by \code{load_vcf_file})}
}
\value{
maximum likelihood tree generated by
  \code{\link[phangorn]{optim.pml}}
}
\description{
Builds maximum likelihood tree using \code{\link[phangorn]{optim.pml}} with a
GTR model of nucleotide substitution.
}
\examples{
ML R code inspired by:
https://www.molecularecologist.com/2016/02/quick-and-dirty-tree-building-in-r/
}
