% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_if_binary_matrix}
\alias{check_if_binary_matrix}
\title{Check that the matrix contains only 1s and/or 0s.}
\usage{
check_if_binary_matrix(mat)
}
\arguments{
\item{mat}{Matrix.}
}
\description{
Function gives error if matrix is not binary; otherwise, returns nothing.
}
