% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_is_number}
\alias{check_is_number}
\title{Check that function input is a numeric.}
\usage{
check_is_number(num)
}
\arguments{
\item{num}{Number}
}
\description{
Doesn't return anything. Gives an error message if input is not a number.
}
\examples{
\dontrun{
check_is_number(100)
}
}
