% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_is_this_class}
\alias{check_is_this_class}
\title{Checks if an object is of the expected R class.}
\usage{
check_is_this_class(obj, current_class)
}
\arguments{
\item{obj}{Any R object.}

\item{current_class}{Character string. Name of R class}
}
\description{
Doesn't return anything. Gives error if the object is not of the expected R
class.
}
\examples{
object <- "example"
check_is_this_class(object, "character")
}
