% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_is_tree}
\alias{check_is_tree}
\title{Check that the input tree is actually a 'phylo' object.}
\usage{
check_is_tree(tree)
}
\arguments{
\item{tree}{Phylogenetic tree.}
}
\description{
Doesn't return anything. Gives an error message if the object is not a
'phylo' object.
}
\examples{
\dontrun{
tree <- ape::rtree(10)
check_is_tree(tree)
}
}
