% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_setequal_tree_mat}
\alias{check_setequal_tree_mat}
\title{Confirm that the tree and variant matrix contain exactly the same samples}
\usage{
check_setequal_tree_mat(tip_labels, colnames_mat)
}
\arguments{
\item{tip_labels}{Character. Vector of tree$tip.labels.}

\item{colnames_mat}{Character. Vector of column names from variant matrix.}
}
\description{
Doesn't return anything. Gives error if the two inputs do not match.
}
