% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{check_tree_is_rooted}
\alias{check_tree_is_rooted}
\title{Check that the tree has a root.}
\usage{
check_tree_is_rooted(tree)
}
\arguments{
\item{tree}{Phylogenetic tree.}
}
\description{
Doesn't return anything. Gives an error message if the object is not a
rooted tree.
}
\examples{
tree <- ape::rtree(10)
check_tree_is_rooted(tree)
}
