% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{clean_up_cds_name_from_gff}
\alias{clean_up_cds_name_from_gff}
\title{Remove ID prefix from CDS feature descriptions (e.g. gene names).}
\usage{
clean_up_cds_name_from_gff(gff)
}
\arguments{
\item{gff}{Data.frame. Rows = genomic features.}
}
\value{
gff: Data.frame.
}
\description{
Remove ID prefix from CDS feature descriptions (e.g. gene names).
}
