% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_snps_by_gene.R
\name{collapse_snps_into_genes}
\alias{collapse_snps_into_genes}
\title{Collapse SNPs into the gene(s) which they are from}
\usage{
collapse_snps_into_genes(bin_mat, gene_vec)
}
\arguments{
\item{bin_mat}{Matrix.}

\item{gene_vec}{Character. Vector of gene names.}
}
\value{
gene_mat: Matrix.
}
\description{
Collapse SNPs into the gene(s) which they are from
}
