% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_snps_by_gene.R
\name{collapse_snps_into_genes_by_impact}
\alias{collapse_snps_into_genes_by_impact}
\title{Collapse SNPs into the gene(s) which they are from and by snpeff impcat}
\usage{
collapse_snps_into_genes_by_impact(
  bin_mat,
  gene_vec,
  predicted_impact,
  snpeff_grouping
)
}
\arguments{
\item{bin_mat}{Matrix.}

\item{gene_vec}{Character. Vector of gene names.}

\item{predicted_impact}{Character. Vector of predicted functional impacts.}

\item{snpeff_grouping}{Character. Vector or single string of impacts of
interest.}
}
\value{
a list of gene_mats, collapsed by gene and snpeff impact
}
\description{
Collapse SNPs into the gene(s) which they are from and by snpeff impcat
}
