% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_snps_in_overlapping_genes.R
\name{dup_snps_in_overlapping_genes}
\alias{dup_snps_in_overlapping_genes}
\title{Duplicate SNPs found in overlapping gene pairs.}
\usage{
dup_snps_in_overlapping_genes(bin_mat, gff_mat)
}
\arguments{
\item{bin_mat}{Matrix. Binary matrix.}

\item{gff_mat}{Matrix. GFF information.}
}
\value{
bin_mat_dup: Matrix.
}
\description{
Duplicate SNPs found in overlapping gene pairs.
}
