% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{format_inputs}
\alias{format_inputs}
\title{Format user-provided inputs to prewas().}
\usage{
format_inputs(dna, tree, outgroup = NULL, gff = NULL)
}
\arguments{
\item{dna}{Character. Path to VCF file containing variant information.}

\item{tree}{NULL, character, or phylo. If character it should be a path to
the tree file.}

\item{outgroup}{NULL or character. If character it should be either a string
naming the outgroup in the dataset or a path to a file containing the
outgroup name.}

\item{gff}{NULL or character. If character it should be a path to the GFF
file.}
}
\value{
A list with the following four elements:
  \describe{
    \item{dna}{vcfR}
    \item{tree}{phylo}
    \item{outgroup}{character}
    \item{gff}{data.frame}
  }
}
\description{
Format user-provided inputs to prewas().
}
