% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_test_dna}
\alias{generate_test_dna}
\title{Given a phylogenetic tree generate plausible DNA sequences for each tip.}
\usage{
generate_test_dna(tree, seq_length, seed)
}
\arguments{
\item{tree}{Phylogenetic tree.}

\item{seq_length}{Number. Length of the generated DNA sequence.}

\item{seed}{Number. Seed for set.seed() function.}
}
\value{
dna_seq: phyDat.
}
\description{
Given a phylogenetic tree generate plausible DNA sequences for each tip.
}
