% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_test_gff}
\alias{generate_test_gff}
\title{Given a dna sequence, generate a dummy GFF3 formatted file.}
\usage{
generate_test_gff(phydat_obj, seq_length)
}
\arguments{
\item{phydat_obj}{phyDat object. DNA alignment.}
}
\description{
Given a dna sequence, generate a dummy GFF3 formatted file.
}
