% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_test_trees}
\alias{generate_test_trees}
\title{Generate multiple phylogenetic trees to use as test data for package.}
\usage{
generate_test_trees(num_samples, seed)
}
\arguments{
\item{seed}{Integer to act as seed for random tree generation process.}
}
\description{
Generate multiple phylogenetic trees to use as test data for package.
}
