% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_alleles.R
\name{get_ancestral_alleles}
\alias{get_ancestral_alleles}
\title{Get ancestral state of alleles}
\usage{
get_ancestral_alleles(tree, mat)
}
\arguments{
\item{tree}{(\code{ape phylo}). Rooted tree.}

\item{mat}{Matrix. Allele matrix. Rows are variants, columns are samples.}
}
\value{
list of two elements:
  \describe{
    \item{ar_results}{Data.frame. Data.frame of most likely ancestral allele
    for each row in allele matrix and probability that that is the ancestral
    state.}
    \item{tree}{phylo. Tree used for ancestral state reconstruction.}
  }
}
\description{
Finds the most likely ancestral allele for each variant position in an allele
matrix using \code{\link[ape]{ace}}, given a rooted tree. Returns the most
likely ancestral allele, its probability, and the tree used to perform
ancestral reconstruction.
}
