% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_snps_by_gene.R
\name{get_gene_names}
\alias{get_gene_names}
\title{Get names of genes from a binary allele matrix}
\usage{
get_gene_names(bin_mat)
}
\arguments{
\item{bin_mat}{Matrix.}
}
\value{
gene_names: Character. Vector of characters. Length = nrow(bin_mat).
}
\description{
Get names of genes from a binary allele matrix
}
