% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_alleles.R
\name{get_major_alleles}
\alias{get_major_alleles}
\title{Get_major_alleles}
\usage{
get_major_alleles(allele_mat)
}
\arguments{
\item{allele_mat}{allele matrix}
}
\value{
major allele for each position
}
\description{
Get_major_alleles
}
