% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gff}
\alias{gff}
\title{GFF3 file for example genomes}
\format{
A character matrix with 110 rows and 9 columns:
\describe{
  \item{Column 1}{Chromosome}
  \item{Column 2}{Data source}
  \item{Column 3}{Feature type}
  \item{Column 4}{Feature start position}
  \item{Column 5}{Feature stop position}
  \item{Column 6}{Score}
  \item{Column 7}{Strand}
  \item{Column 8}{Phase}
  \item{Column 9}{Locus ID}
 }
}
\usage{
gff
}
\description{
An example of GFF3 formatted genome information.
}
\keyword{datasets}
