% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_allele_matrix.R
\name{identify_variant_sites}
\alias{identify_variant_sites}
\title{Find all rows in allele matrix with SNPs.}
\usage{
identify_variant_sites(mat)
}
\arguments{
\item{mat}{Matrix.}
}
\value{
rows_to_keep: Logical vector.
}
\description{
Give an error if there are no variant sites in the dataset.
}
