% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_this_class}
\alias{is_this_class}
\title{Tests if an input object has the specified class.}
\usage{
is_this_class(obj, current_class)
}
\arguments{
\item{obj}{Any R object.}

\item{current_class}{String. Name of the expected class of the R object.}
}
\value{
is_this_class: Logical.
}
\description{
Tests if an input object has the specified class.
}
\examples{
object <- "example"
is_this_class(object, "character")
}
