% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_allele_matrix.R
\name{keep_only_variant_sites}
\alias{keep_only_variant_sites}
\title{Given a DNA matrix keep only those rows with SNPs.}
\usage{
keep_only_variant_sites(dna_mat)
}
\arguments{
\item{dna_mat}{Matrix.}
}
\value{
variant_only_dna_mat: Matrix.
}
\description{
Given a DNA matrix keep only those rows with SNPs.
}
