% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_alleles.R
\name{make_all_tree_edges_positive}
\alias{make_all_tree_edges_positive}
\title{Make tree edges positive}
\usage{
make_all_tree_edges_positive(tree)
}
\arguments{
\item{tree}{(\code{ape phylo}).}
}
\value{
tree: (\code{ape phylo}). Tree now has only positive edge lengths.
}
\description{
If tree edges are zero or negative, makes them a very small positive number
(1/1000th of the smallest edge length). This is to prevent
\code{\link[ape]{ace}} from breaking. A warning is thrown. Returns a tree
with all positive edge lengths.
}
