% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outgroup}
\alias{outgroup}
\title{Name of outgroup in the phylogenetic tree.}
\format{Character string.}
\usage{
outgroup
}
\description{
Name of outgroup in the phylogenetic tree.
}
\keyword{datasets}
