% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{read_gff}
\alias{read_gff}
\title{Read in a GFF file given a file path}
\usage{
read_gff(gff_path)
}
\arguments{
\item{gff_path}{Character. Path to GFF file.}
}
\value{
gff: Data.frame.
}
\description{
Read in a GFF file given a file path
}
