% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_tree_and_vcf.R
\name{read_in_tree}
\alias{read_in_tree}
\title{Read in tree}
\usage{
read_in_tree(tree)
}
\arguments{
\item{tree}{Character or or `ape phylo`. If character: path to a tree file.}
}
\value{
tree: `ape phylo`.
}
\description{
If \code{tree} is a string, \code{read_in_tree} reads in the tree as an `ape
phylo` object. \code{read_in_tree} checks to see if the object is an `ape
phylo` object.
}
