% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_allele_matrix.R
\name{remove_invariant_sites}
\alias{remove_invariant_sites}
\title{Remove any invariant sites from allele matrix.}
\usage{
remove_invariant_sites(mat, rows_to_keep)
}
\arguments{
\item{mat}{Matrix.}

\item{rows_to_keep}{Logical. Vector of TRUE/FALSE. FALSE corresponds to rows
that should be dropped (invariant loci). TRUE corresponds to rows to keep
(variant loci).}
}
\value{
mat: Matrix.
}
\description{
Remove any invariant sites from allele matrix.
}
