% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_alleles.R
\name{remove_unknown_anc_state}
\alias{remove_unknown_anc_state}
\title{Remove unknown ancestral states}
\usage{
remove_unknown_anc_state(allele_mat, ar_results)
}
\arguments{
\item{anc}{}
}
\description{
Remove rows from variant matrix where the ancestral state is unknown (- or N)
}
