% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_allele_matrix.R
\name{replace_non_ATGC_with_N}
\alias{replace_non_ATGC_with_N}
\title{In allele matrix replace non-standard alleles with N}
\usage{
replace_non_ATGC_with_N(mat)
}
\arguments{
\item{mat}{Matrix.}
}
\value{
mat: Matrix.
}
\description{
In allele matrix replace non-standard alleles with N
}
