% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{save_gff3}
\alias{save_gff3}
\title{Save GFF file}
\usage{
save_gff3(gff, file_prefix)
}
\arguments{
\item{gff}{Matrix.}

\item{file_prefix}{Character. File prefix (without .gff)}
}
\description{
Doesn't return anything to the R console.
}
