% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{save_tree}
\alias{save_tree}
\title{Save a tree to a .tree file.}
\usage{
save_tree(tree, file_prefix)
}
\arguments{
\item{tree}{Phylogenetic tree.}

\item{file_prefix}{String. File name (without .fna).}
}
\description{
Save a tree to a .tree file.
}
