% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{subset_gff}
\alias{subset_gff}
\title{Keep only GFF features that are in CDS regions.}
\usage{
subset_gff(gff)
}
\arguments{
\item{gff}{Data.frame. Rows = genomic features.}
}
\value{
gff: Data.frame.
}
\description{
Keep only GFF features that are in CDS regions.
}
