% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_tree_and_vcf.R
\name{subset_tree_and_matrix}
\alias{subset_tree_and_matrix}
\title{Subset tree and matrix}
\usage{
subset_tree_and_matrix(tree, mat)
}
\arguments{
\item{tree}{path to tree file or ape phylo object}

\item{mat}{variant matrix (columns are samples, rows are variants)}
}
\value{
List with two objects:
  \describe{
    \item{tree}{\code{ape phylo}.}
    \item{mat}{Matrix.}
  }
}
\description{
Subsets the tree and variant matrix to include only samples in both. Returns
the subsetted tree and matrix.
}
