% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Phylogenetic tree of example genomes}
\format{
An ape phylo object with 14 tips.
}
\usage{
tree
}
\description{
Example unrooted phylogenetic tree.
}
\keyword{datasets}
