% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vcf}
\alias{vcf}
\title{Nucleotide variants in example genome samples}
\format{vcfR class object with three sections:
\describe{
  \item{meta}{The metadata for the VCF file including the file format version
  number}
  \item{fix}{A character matrix with 313 rows and 8 columns. Contains
  information on chromosome (CHROM), genome position (POS), reference genome
  allele (REF), and alternative allele (ALT)}
  \item{gt}{A character matrix with 313 rows and 14 columns. Presence/absence
  for each variant defined in fix. Colnames are sample IDs.}
 }}
\usage{
vcf
}
\description{
An example dataset containing 313 variants from 13 genome samples.
}
\keyword{datasets}
