% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_snps_in_overlapping_genes.R
\name{dup_snps_in_overlapping_genes}
\alias{dup_snps_in_overlapping_genes}
\title{Title}
\usage{
dup_snps_in_overlapping_genes(bin_mat, gff_mat)
}
\arguments{
\item{gff_mat}{}
}
\description{
Title
}
