% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_inputs.R
\name{format_inputs}
\alias{format_inputs}
\title{Format user-provided inputs to prewas().}
\usage{
format_inputs(dna, tree = NULL, outgroup = NULL, gff = NULL,
  anc = TRUE)
}
\arguments{
\item{dna}{Character or vcfR. Path to VCF file or vcfR object.}

\item{tree}{NULL, character, or phylo. If character it should be a path to
the tree file.}

\item{outgroup}{NULL or character. If character it should be either a string
naming the outgroup in the dataset or a path to a file containing the
outgroup name.}

\item{gff}{NULL, character, or matrix. If character it should be a path to
the GFF file. If NULL ignored. If a matrix or data.frame it should be the
GFF information stored in 9 columns with the genes as rows.}

\item{anc}{logical. If TRUE ancestral reconstruction is performed and the
ancestral allele is used for referencing. If FALSE no ancestral
reconstruction is performed and the major allele is used for referencing.}
}
\value{
A list with the following four elements:
  \describe{
    \item{dna}{vcfR}
    \item{tree}{phylo}
    \item{outgroup}{character}
    \item{gff}{data.frame}
  }
}
\description{
Format user-provided inputs to prewas().
}
