% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate all of the data necessary to test functionality of prewas package.}
\usage{
generate_test_data(num_samples, seq_length, seed)
}
\arguments{
\item{seed}{}
}
\description{
Generate all of the data necessary to test functionality of prewas package.
}
\examples{
generate_test_data(15, 1000, 1)
}
