% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_alleles.R
\name{get_ancestral_alleles}
\alias{get_ancestral_alleles}
\title{Get ancestral state of alleles}
\usage{
get_ancestral_alleles(tree, mat)
}
\arguments{
\item{tree}{rooted tree}

\item{mat}{allele matrix (rows are variants, columns are samples)}
}
\value{
list of (1) ar_results: matrix of most likely ancestral allele for each row in allele matrix and probability that that is the ancestral state and (2) tree: tree used for ancestral state reconstruction
}
\description{
Rereference alleles based on rooted tree
}
