% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_alleles.R
\name{get_major_alleles}
\alias{get_major_alleles}
\title{Get major alleles}
\usage{
get_major_alleles(allele_mat)
}
\arguments{
\item{allele_mat}{Matrix. Allele matrix.}
}
\value{
major_allele: major allele for each position.
}
\description{
Finds the major allele (most common allele) for each variant position in an
allele matrix.
}
