% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_alleles.R
\name{get_major_alleles}
\alias{get_major_alleles}
\title{Get major alleles}
\usage{
get_major_alleles(allele_mat)
}
\arguments{
\item{allele_mat}{Matrix. Allele matrix. Rows are variants. Columns are
samples.}
}
\value{
major_allele: Character vector. Gives the major allele for each
  position. Names are the genomic loci. Values are the nucleotides.
}
\description{
Finds the major allele (most common allele) for each variant position in an
allele matrix.
}
