% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{load_vcf_file}
\alias{load_vcf_file}
\title{Read in variant data from a VCF file}
\usage{
load_vcf_file(vcf_path)
}
\arguments{
\item{vcf_path}{Character. Path to VCF file.}
}
\value{
vcf_geno_mat: Matrix.
}
\description{
Read in variant data from a VCF file
}
