% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_alleles.R
\name{make_binary_matrix}
\alias{make_binary_matrix}
\title{Make binary matrix from allele matrix}
\usage{
make_binary_matrix(allele_mat, reference_allele)
}
\arguments{
\item{allele_mat}{allele matrix (split by multi-allelic site)}

\item{reference_allele}{vector of alleles that are 0 in binary matrix (ancestral allele or major allele)}
}
\description{
Make binary matrix from allele matrix
}
