% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{read_gff}
\alias{read_gff}
\title{Read in a GFF file given a file path}
\usage{
read_gff(gff)
}
\arguments{
\item{gff}{Character or matrix. If character: path to GFF file.}
}
\value{
gff: matrix
}
\description{
Read in a GFF file given a file path
}
