% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_tree_and_fasta.R
\name{read_in_tree}
\alias{read_in_tree}
\title{Read in tree}
\usage{
read_in_tree(tree)
}
\arguments{
\item{tree}{tree path or ape phylo object}
}
\value{
ape phylo object
}
\description{
Read in tree
}
