% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_alleles.R
\name{remove_unknown_alleles}
\alias{remove_unknown_alleles}
\title{Remove unknown ancestral states}
\usage{
remove_unknown_alleles(allele_mat, alleles, ar_results)
}
\arguments{
\item{allele_mat}{Matrix. Allele matrix. Rows are variants, columns are
samples.}

\item{alleles}{Factor. Vector of reference alleles (ancestral alleles or
major alleles)}

\item{ar_results}{Data.frame. Results from ancestral reconstruction}
}
\value{
A list of three objects:
  \describe{
    \item{allele_mat}{Matrix. Rows are variants. Columns are samples.}
    \item{ar_results}{Data.frame. Variants with unknown ancestral states
    removed. Rows are variants. If ancestral reconstruction was performed:
    first column in ancestral allele & second column is probability. If no
    ancestral reconstruction performed: the only column is the major allele.}
    \item{removed}{Character. Vector with names of removed samples.}
  }
}
\description{
Removes rows from variant matrix where the reference allele (ancestral allele
or major allele) is unknown (- or N).
}
