% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_alleles.R
\name{remove_unknown_alleles}
\alias{remove_unknown_alleles}
\title{Remove unknown ancestral states}
\usage{
remove_unknown_alleles(allele_mat, alleles, ar_results)
}
\arguments{
\item{ar_results}{}
}
\description{
Remove rows from variant matrix where the ancestral state is unknown (- or N)
}
