% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_tree_and_vcf.R
\name{root_tree}
\alias{root_tree}
\title{Root tree on outgroup}
\usage{
root_tree(tree, outgroup = NULL)
}
\arguments{
\item{tree}{Character or phylo. If character: path to a tree file.}

\item{outgroup}{Character. Tip name of outgroup in phylogeny. If NULL,
midpoint root if not rooted}
}
\value{
tree: phylo. Rooted tree without outgroup.
}
\description{
Roots tree based on an outgroup. If the outgroup is provided, roots based on
the outgroup. If the outgroup is null and the tree isn't rooted, midpoint
roots the tree. If the tree is rooted and no outgroup is provided, returns
the tree as is.
}
\examples{
\dontrun{
tree <- ape::rcoal(100)
ape::is.rooted(tree)
tree <- root_tree(tree)
ape::is.rooted(tree)
}
}
