% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_tree_and_fasta.R
\name{root_tree}
\alias{root_tree}
\title{Root tree on outgroup}
\usage{
root_tree(tree, outgroup = NULL)
}
\arguments{
\item{tree}{phylogenetic tree or file path of tree}

\item{outgroup}{tip name of outgroup in phylogeny. If NULL, midpoint root if not rooted}
}
\value{
rooted tree without outgroup
}
\description{
Root tree based on outgroup. If outgroup is provided, root based on outgroup. If outgroup is null and the tree isn't rooted, midpoint root tree. If tree is rooted and no outgroup is provided, return tree as is.
}
\examples{
tree = rcoal(100)
is.rooted(tree)
root_tree_og(tree)
is.rooted(tree)
}
