% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{save_fasta}
\alias{save_fasta}
\title{Save a fasta file given a phyDat sequence object.}
\usage{
save_fasta(phydat_obj, file_prefix)
}
\arguments{
\item{phydat_obj}{phyDat. DNA sequence.}

\item{file_prefix}{Character. File name (without .fna).}
}
\description{
Does not return anything in R console.
}
