% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{save_tree}
\alias{save_tree}
\title{Save a tree to a .tree file.}
\usage{
save_tree(tree, file_prefix)
}
\arguments{
\item{tree}{Phylo.}

\item{file_prefix}{Character. File name (without .tree).}
}
\description{
Does not return anything in R console.
}
