% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_multiallelic_snps.R
\name{split_multi_to_biallelic_snps}
\alias{split_multi_to_biallelic_snps}
\title{Single line to multi line representation of multiallelic sites}
\usage{
split_multi_to_biallelic_snps(mat, ar_results)
}
\arguments{
\item{mat}{- matrix}

\item{ar_results}{- data.frame}
}
\value{
mat_split, ar_results_split, split_rows_flag
}
\description{
Changes the single line representation of multiallelic sites to a multi line
representation. That is, rows that have N alleles are replicated N-1 times.
}
